added help flag and split up preprocess in own file.
