+++ cpu/sha_preprocess.hpp
@@ -1,0 +1,6 @@
+#ifndef __SHA_PREPROCESS_H_INCLUDED__
+#define __SHA_PREPROCESS_H_INCLUDED__
+
+void pre_process(char* element);
+
+ #endif
